// src/lib/tesserafin-sdk/generated/models/playback-capability-dto.ts
/* tslint:disable */
/**
 * Tesserafin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PlaybackCapabilityScope } from './playback-capability-scope';

/**
 * A minted playback capability.
 * @export
 * @interface PlaybackCapabilityDto
 */
export interface PlaybackCapabilityDto {
    /**
     * Gets or sets the public identifier. Renewal is addressed by this, not by the secret, so the secret never travels a second time. Safe to log.
     * @type {string}
     * @memberof PlaybackCapabilityDto
     */
    'CapabilityId'?: string;
    /**
     * Gets or sets when the capability stops being accepted.
     * @type {string}
     * @memberof PlaybackCapabilityDto
     */
    'ExpiresAt'?: string;
    /**
     * Gets or sets when the capability was issued.
     * @type {string}
     * @memberof PlaybackCapabilityDto
     */
    'IssuedAt'?: string;
    /**
     * Gets or sets the bound item, when the scopes are item-bound.
     * @type {string}
     * @memberof PlaybackCapabilityDto
     */
    'ItemId'?: string | null;
    /**
     * Gets or sets the bound media source, when the scopes are item-bound.
     * @type {string}
     * @memberof PlaybackCapabilityDto
     */
    'MediaSourceId'?: string | null;
    /**
     * Gets or sets the bound play session.
     * @type {string}
     * @memberof PlaybackCapabilityDto
     */
    'PlaySessionId'?: string;
    /**
     * Gets or sets the kinds of media this capability may fetch.
     * @type {Array<PlaybackCapabilityScope>}
     * @memberof PlaybackCapabilityDto
     */
    'Scopes'?: Array<PlaybackCapabilityScope>;
    /**
     * Gets or sets the secret to present on media requests. Sensitive despite being short-lived: short-lived is not non-sensitive.
     * @type {string}
     * @memberof PlaybackCapabilityDto
     */
    'Value'?: string;
}