// src/lib/tesserafin-sdk/generated/models/playback-capability-renewal-dto.ts
/* tslint:disable */
/**
 * Tesserafin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * The result of extending a capability\'s expiry.
 * @export
 * @interface PlaybackCapabilityRenewalDto
 */
export interface PlaybackCapabilityRenewalDto {
    /**
     * Gets or sets the capability that was renewed.
     * @type {string}
     * @memberof PlaybackCapabilityRenewalDto
     */
    'CapabilityId'?: string;
    /**
     * Gets or sets the new expiry.
     * @type {string}
     * @memberof PlaybackCapabilityRenewalDto
     */
    'ExpiresAt'?: string;
    /**
     * Gets or sets the original issue time, which renewal does not move.
     * @type {string}
     * @memberof PlaybackCapabilityRenewalDto
     */
    'IssuedAt'?: string;
}