// src/lib/tesserafin-sdk/generated/models/playback-capability-request-dto.ts
/* tslint:disable */
/**
 * Tesserafin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PlaybackCapabilityScope } from './playback-capability-scope';

/**
 * Asks the server to mint a playback capability.
 * @export
 * @interface PlaybackCapabilityRequestDto
 */
export interface PlaybackCapabilityRequestDto {
    /**
     * Gets or sets the item whose media this capability may fetch. Omitted only for a scope that is not item-bound, which today means Tesserafin.Controller.Net.PlaybackCredentials.PlaybackCapabilityScope.Fonts.
     * @type {string}
     * @memberof PlaybackCapabilityRequestDto
     */
    'ItemId'?: string | null;
    /**
     * Gets or sets the media source within that item.
     * @type {string}
     * @memberof PlaybackCapabilityRequestDto
     */
    'MediaSourceId'?: string | null;
    /**
     * Gets or sets the play session this capability belongs to. Ending that play session revokes this capability and no other.
     * @type {string}
     * @memberof PlaybackCapabilityRequestDto
     */
    'PlaySessionId': string;
    /**
     * Gets or sets the kinds of media this capability may fetch. Must name at least one; a capability with no scope would grant nothing.
     * @type {Array<PlaybackCapabilityScope>}
     * @memberof PlaybackCapabilityRequestDto
     */
    'Scopes': Array<PlaybackCapabilityScope>;
}