// src/lib/tesserafin-sdk/generated/models/playback-capability-scope.ts
/* tslint:disable */
/**
 * Tesserafin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * The kind of media a playback capability may fetch.
 * @export
 * @enum {string}
 */

export const PlaybackCapabilityScope = {
    Media: 'Media',
    Subtitles: 'Subtitles',
    Attachments: 'Attachments',
    Trickplay: 'Trickplay',
    Fonts: 'Fonts'
} as const;

export type PlaybackCapabilityScope = typeof PlaybackCapabilityScope[keyof typeof PlaybackCapabilityScope];