// src/lib/tesserafin-sdk/generated/models/web-socket-ticket-dto.ts
/* tslint:disable */
/**
 * Tesserafin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * A minted single-use WebSocket ticket.
 * @export
 * @interface WebSocketTicketDto
 */
export interface WebSocketTicketDto {
    /**
     * Gets or sets when the ticket stops being accepted. Seconds after issue, not minutes.
     * @type {string}
     * @memberof WebSocketTicketDto
     */
    'ExpiresAt'?: string;
    /**
     * Gets or sets when the ticket was issued.
     * @type {string}
     * @memberof WebSocketTicketDto
     */
    'IssuedAt'?: string;
    /**
     * Gets or sets the public identifier. Safe to log.
     * @type {string}
     * @memberof WebSocketTicketDto
     */
    'TicketId'?: string;
    /**
     * Gets or sets the secret to present on the upgrade request.
     * @type {string}
     * @memberof WebSocketTicketDto
     */
    'Value'?: string;
}